split config topic to files
