+++ cmd/kafeman/topic_cmd/topics.go
@@ -1,0 +1,10 @@
+package topic_cmd
+
+import "github.com/spf13/cobra"
+
+func NewTopicsCMD() *cobra.Command {
+	cmd := NewLSTopicsCMD()
+	cmd.Use = "topics"
+
+	return cmd
+}
